style: 파일분리 및 폴더정리
