+++ OpenMarket/OpenMarket/Model/URLSession/URLSessionProvider.swift
@@ -7,27 +7,6 @@
 
 import Foundation
 
-enum Request {
-    case healthChecker
-    case productList(pageNumber: Int, itemsPerPage: Int)
-    case productDetail(productNumber: Int)
-    
-    var url: URL? {
-        switch self {
-        case .healthChecker:
-            return URLComponents.healthCheckUrl
-        case .productList(let pageNumber, let itemsPerPage):
-            return URLComponents.marketUrl(path: nil,
-                                           queryItems: [URLQueryItem(name: "page_no",
-                                                                     value: String(pageNumber)),
-                                                        URLQueryItem(name: "items_per_page",
-                                                                     value: String(itemsPerPage))])
-        case .productDetail(let productNumber):
-            return URLComponents.marketUrl(path: [String(productNumber)], queryItems: nil)
-        }
-    }
-}
-
 class MarketURLSessionProvider {
     let session: URLSessionProtocol
     var market: Market?
